refactor sync service and clean up code
